% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comarcas_tjmg.R
\name{comarcas_tjmg}
\alias{comarcas_tjmg}
\title{Consolida a lista de municipios do TJMG com o IBGE, retornando as comarcas dos municipios}
\usage{
comarcas_tjmg()
}
\value{
Data frame consolidado dos munícipios e comarcas do TJMG, incluindo informações do IBGE
}
\description{
Consolida a lista de municipios do TJMG com o IBGE, retornando as comarcas dos municipios
}
