% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_municipios_ibge.R
\name{read_municipios_ibge}
\alias{read_municipios_ibge}
\title{Funcao que baixa o arquivo Excel contendo a listagem oficial de municipios cadastrados no IBGE.}
\usage{
read_municipios_ibge(uf = "MG")
}
\arguments{
\item{uf}{Sigla da Unidade Federativa a ser baixada}
}
\value{
Data frame dos municípios do IBGE por UF selecionada
}
\description{
Funcao que baixa o arquivo Excel contendo a listagem oficial de municipios cadastrados no IBGE.
}
