% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tjmg.R
\name{read_tjmg}
\alias{read_tjmg}
\title{Funcao que baixa o pdf das comarcas do TJMG, trata e cria um data frame com municipios, distritos e suas comarcas}
\usage{
read_tjmg()
}
\value{
Data frame dos municípios e distritos de MG, bem como as respectivas comarcas do TJMG
}
\description{
Funcao que baixa o pdf das comarcas do TJMG, trata e cria um data frame com municipios, distritos e suas comarcas
}
